+++ lib/client.js
@@ -4,7 +4,7 @@ var decorator = require("./decorator"),
     _ = require("./utils");
 
 var protocols = {
-  http: require("./drivers/http")
+  http: require("./protocols/http")
 };
 
 function checkArgs(protocol, opts) {
